+++ internal/core/repolint/serialize.go
@@ -1,4 +1,4 @@
-package audit
+package repolint
 
 import "encoding/json"